don't die if buffer->segment conversion fails (bz://504)
Buffer->segment conversion now takes place in a separate, long-running process.  This
means that two conversions will never happen at the same time.  The process is restarted
if it dies, and all buffers except the current one are re-added to its queue.

The biggest danger here is that buffer_converter/2 ends up in a tight-failure loop, being
restarted constantly because a bad Buffer is repeatedly sent to it.  This process should
be moved into its own supervisor-monitored gen_server for better restart semantics.

The final call back to the mi_server has also been changed to a cast, just like
compaction, since that process doesn't need to wait on the result of that message.

Also moved ".deleted" into a macro, so I wouldn't mistype it.