remove outdated comments (bz://504)
Answers to TODOs are:

Q: How does caller know if worker proc has crashed?
A: The exit of the worker is now trapped by the mi_server.  An end-of-table message is
sent to the caller.

Q: If filter fun or other crashes, server crashes.
A: This is no longer the case.  Exits of sub-processes are now trapped, and the
subprocesses are either restarted (for compaction and buffer->segment conversion) or
cleaned up (for stream/range queries).