Check for race in buffer_to_segment
tags: az380

A race condition exists between the `merge_idex:dropq call and the
`mi_buffer_converter:convert` call.  Since all convert calls are done
async there are cases where `mi_buffer_converter` will be working with
old buffers that have since been deleted and since these data
strcutures are backed by public ETS tables it means `badarg`s will
ensue.  There are two races that can occur:

1. `mi_buffer_converter` performs an operation on a buffer or segment
that no longer exists and crashes.  I'm leaving this case alone as the
supervisor will restart the worker.  The only thing I changed he was
the maximum restarts from 2 to 10.

2. `mi_buffer_converter` calls `mi_server:buffer_to_segment` with a
since deleted buffer (or with one about to be deleted because a drop
call is in the mi_server mailbox) and mi_server crashes while handling
`buffer_to_segment`.  This would potentially crash whatever it is
linked to and I thought a better behavior would be to instead ignore
request for non-existant buffers and log a warning.

To be clear I'm defining the semantics of drop to mean "clear/reset
all my data" and I don't think this should result in a crash in the
case where `mi_buffer_converter` is working with old data.