Don't crash b/c of stale buffer
tags: az380

buffer_converter doesn't play well with drop b/c of it's async behavior.
I opted to modify it to simply log a warning msg in the case of a
non-existent buffer since a drop call is the most likely cause.  Without
this change running large iterations of QC causes the supervisor to crash
which then causes QC to go into an infinite loop trying to shrink the failing
case but that will never work because it only manifests itself in the bigger
test case.