Merge in `lookup` and `range`
tags: az296

The `stream` API has now been renamed to `lookup` and `lookup_sync`
returning an iterator and list respectively.  There was a similar
change to range but it kept it's name.  Instead of passing a Pid to
receive the data the worker process now feeds data directly to the
caller.

Furthermore, the filter application was pushed down to clean up the
code a bit **and** avoid an additional traversal of the result set.