Add an explicit error return to lookup/range
tags: az442

If an error occurs during lookup/range the current behavior is to
return an `eof`, thus truncating the data and leaving the caller
blissfully unaware that an error occured.  In the case of riak_search
this means if an error occurs during a lookup/range then the client will
think it is working with the full set of results when it is not.

This patch changes the lookup/range API to instead explicitly return an
`{error, Reason}` tuple so that the client is made aware that an
error has occured.