Add back pressure to iterator API
In order to prevent slower consumers from piling up result chunks add
back pressure to the iterator API.  This is important because this API
is used for handoff & repair both of which have back pressure and will
potentially not be able to consume data as fast as the local iterator
can produce it.

Currently after every 4 result chunks sent the MI server iterator will
wait until it sees a continue msg.  The number 4 was chosen
arbitrarily and the reason for not applying back pressure on a
per-chunk basis is to avoid some latency/overhead by reducing the
number of msgs that need to be sent.