Fix iterator API
The iterator API will timeout any time there is a segment larger than
the `segment_full_read_size`.  In this scenario the segment file is
opened by the `mi_server` process but then reading is attempted by the
spawned iterator process.  Since the segment file is opened in raw mode
this will fail.  The port message will go to the `mi_server` causing
an `Unexpected info...` msg and the iteraor will hit the
`lookup_timeout`.

Many thanks to Arnaud Wetzel who discovered this bug and took time to
explain it to me.