Add segment .data file corruption test
Write some entries to a buffer.  Convert it to a segment.  Replace one
byte of the .data file with a random byte.  Open the segment.  Try to
iterate over it.

If doing a git bisect this commit will fail as the corruption
resilience code has not yet been added for this scenario.