Check for mutated term, close FH on error
Since merge index doesn't have CRCs there is the potential (albeit
extremly rare in most production scenarios) for bytes to be corrupted
in such a way that binary_to_term doesn't fail but the term itself has
been mutated.  E.g. the arity of a tuple, changing a list to an atom,
etc.