Remove the delayed_write option
Don't use the delayed_write option when opening a file.  This option
is known to cause corruption issues in Erlang 15x.  It is fixed in
16B01 but currently Riak must run on 15B01, thus the VM cannot simply
be upgraded.

This patch is being made in direct relation to a customer having
buffer corruption issues when the machine is under heavy load.  I was
able to reproduce easily using an underpowered centos VM running on
virtual box.