Remove drop call from EQC test
The drop call causes occasional errors with the EQC test.  I'm
removing it because the EQC test does not model how Riak Search uses
the drop call.  Which is only done after a handoff has finished.  It
would be nice to find and fix the race condition that exists between
drop and the other operations but for now I just want a consistent
test.